consume book service
