+++ ng5/src/app/app-routing.module.ts
@@ -8,7 +8,7 @@ import { AuthComponent } from 'src/app/screens/auth/auth.component';
 import { BookListComponent } from 'src/app/screens/auth/screens/book-list/book-list.component';
 import { UnauthGuard } from './authorization/unauth/unauth.guard';
 import { InitialFormComponent } from 'src/app/screens/unauth/screens/initial-form/initial-form.component';
-import { BookDetailComponent } from '../app/screens/auth/screens/book-list/screens/book-detail/book-detail.component';
+import { BookDetailComponent } from './screens/auth/screens/book-detail/book-detail.component';
