+++ ng5/src/app/screens/auth/screens/book-detail/book-detail.component.spec.ts
