--- ng5/src/app/screens/auth/screens/book-list/screens/book-detail/book-detail.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-book-detail',
-  templateUrl: './book-detail.component.html',
-  styleUrls: ['./book-detail.component.scss']
-})
-export class BookDetailComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
